+++ chromadb/segment/impl/manager/local.py
@@ -24,7 +24,7 @@ class SegmentType(Enum):
 SEGMENT_TYPE_IMPLS = {
     SegmentType.SQLITE: "chromadb.segment.impl.metadata.sqlite.SqliteMetadataSegment",
     SegmentType.HNSW_LOCAL_MEMORY: "chromadb.segment.impl.vector.local_hnsw.LocalHnswSegment",
-    SegmentType.HNSW_LOCAL_PERSISTED: "chromadb.segment.impl.vector.local_hnsw.PersistentLocalHnswSegment",
+    SegmentType.HNSW_LOCAL_PERSISTED: "chromadb.segment.impl.vector.local_persistent_hnsw.PersistentLocalHnswSegment",
 }
