+++ chromadb/segment/impl/vector/local_hnsw.py
@@ -1,9 +1,10 @@
 from overrides import override
-from typing import Any, Optional, Sequence, Dict, Set, List, Callable, Union, cast
+from typing import Optional, Sequence, Dict, Set, List, Callable, Union, cast
 from uuid import UUID
 from chromadb.segment import VectorReader
 from chromadb.ingest import Consumer
 from chromadb.config import System, Settings
+from chromadb.segment.impl.vector.batch import Batch
 from chromadb.types import (
     EmbeddingRecord,
     VectorEmbeddingRecord,
@@ -21,12 +22,6 @@ import multiprocessing
 import hnswlib
 from threading import Lock
 import logging
-import os
-import pickle
-import numpy as np
-import numpy.typing as npt
-
-from chromadb.utils import distance_functions
 
 logger = logging.getLogger(__name__)
 
@@ -62,256 +57,6 @@ class HnswParams:
             metadata.get("hnsw:num_threads", multiprocessing.cpu_count())
         )
         self.resize_factor = float(metadata.get("hnsw:resize_factor", 1.2))
-
-
-class Batch:
-    """Used to model the set of changes as an atomic operation"""
-
-    _ids_to_records: Dict[str, EmbeddingRecord]
-    _deleted_ids: Set[str]
-    _written_ids: Set[str]
-    _upsert_add_ids: Set[str]  # IDs that are being added in an upsert
-    add_count: int
-    update_count: int
-    delete_count: int
-    max_seq_id: SeqId
-
-    def __init__(self) -> None:
-        self._ids_to_records = {}
-        self._deleted_ids = set()
-        self._written_ids = set()
-        self._upsert_add_ids = set()
-        self.add_count = 0
-        self.update_count = 0
-        self.delete_count = 0
-        self.max_seq_id = 0
-
-    def __len__(self) -> int:
-        """Get the number of changes in this batch"""
-        return len(self._written_ids) + len(self._deleted_ids)
-
-    def count(self) -> int:
-        """Get the net number of embeddings in this batch"""
-        return len(self._written_ids)
-
-    def get_deleted_ids(self) -> List[str]:
-        """Get the list of deleted embeddings in this batch"""
-        return list(self._deleted_ids)
-
-    def get_written_ids(self) -> List[str]:
-        """Get the list of written embeddings in this batch"""
-        return list(self._written_ids)
-
-    def get_written_vectors(self) -> List[Vector]:
-        """Get the list of vectors to write in this batch"""
-        return [
-            cast(Vector, self._ids_to_records[id]["embedding"])
-            for id in self.get_written_ids()
-        ]
-
-    def get_record(self, id: str) -> EmbeddingRecord:
-        """Get the record for a given ID"""
-        return self._ids_to_records[id]
-
-    def is_new(self, id: str) -> bool:
-        """Returns true if the id is a new addition to the index"""
-        record = self._ids_to_records[id]
-        return record["operation"] == Operation.ADD or (
-            record["operation"] == Operation.UPSERT and id in self._upsert_add_ids
-        )
-
-    def apply(self, record: EmbeddingRecord, is_add: bool = False) -> None:
-        """
-        Apply an embedding record to this batch. Records passed to this method are assumed to be validated for correctness.
-        For example, a delete or update presumes the ID exists in the index. An add presumes the ID does not exist in the index.
-        In the case of upsert, the is_add flag should be set to True if the ID does not exist in the index, and False otherwise.
-        """
-
-        id = record["id"]
-        if record["operation"] == Operation.DELETE:
-            self._deleted_ids.add(id)
-
-            # If the ID was previously written, remove it from the written set
-            # And update the add/update/delete counts
-            if id in self._written_ids:
-                self._written_ids.remove(id)
-                if self._ids_to_records[id]["operation"] == Operation.ADD:
-                    self.add_count -= 1
-                elif self._ids_to_records[id]["operation"] == Operation.UPDATE:
-                    self.update_count -= 1
-                elif self._ids_to_records[id]["operation"] == Operation.UPSERT:
-                    if id in self._upsert_add_ids:
-                        self.add_count -= 1
-                        self._upsert_add_ids.remove(id)
-                    else:
-                        self.update_count -= 1
-
-            # Remove the record from the batch
-            if id in self._ids_to_records:
-                del self._ids_to_records[id]
-
-            self.delete_count += 1
-        else:
-            self._ids_to_records[id] = record
-            self._written_ids.add(id)
-
-            # If the ID was previously deleted, remove it from the deleted set
-            # And update the delete count
-            if id in self._deleted_ids:
-                self._deleted_ids.remove(id)
-                self.delete_count -= 1
-
-            # Update the add/update counts
-            if record["operation"] == Operation.UPSERT:
-                if is_add:
-                    self.add_count += 1
-                    self._upsert_add_ids.add(id)
-                else:
-                    self.update_count += 1
-            elif record["operation"] == Operation.ADD:
-                self.add_count += 1
-            elif record["operation"] == Operation.UPDATE:
-                self.update_count += 1
-
-        self.max_seq_id = max(self.max_seq_id, record["seq_id"])
-
-
-class BruteForceIndex:
-    """A lightweight, numpy based brute force index that is used for batches that have not been indexed into hnsw yet"""
-
-    # TODO: mark internal
-    id_to_index: Dict[str, int]
-    index_to_id: Dict[int, str]
-    id_to_seq_id: Dict[str, int]
-    deleted_ids: Set[str]
-    curr_index: int
-    size: int
-    dimensionality: int
-    distance_fn: Callable[[npt.NDArray[Any], npt.NDArray[Any]], float]
-    vectors: npt.NDArray[Any]
-
-    def __init__(self, size: int, dimensionality: int, space: str = "l2"):
-        if space == "l2":
-            self.distance_fn = distance_functions.l2
-        elif space == "ip":
-            self.distance_fn = distance_functions.ip
-        elif space == "cosine":
-            self.distance_fn = distance_functions.cosine
-        else:
-            raise Exception(f"Unknown distance function: {space}")
-
-        self.id_to_index = {}
-        self.index_to_id = {}
-        self.id_to_seq_id = {}
-        self.deleted_ids = set()
-        self.curr_index = 0
-        self.size = size
-        self.dimensionality = dimensionality
-        self.vectors = np.zeros((size, dimensionality))
-
-    def __len__(self) -> int:
-        return len(self.id_to_index)
-
-    def flush(self) -> None:
-        self.id_to_index = {}
-        self.index_to_id = {}
-        self.id_to_seq_id = {}
-        self.deleted_ids.clear()
-        self.curr_index = 0
-        self.vectors.fill(0)
-
-    # TODO: thread safety
-    def upsert(self, records: List[EmbeddingRecord]) -> None:
-        if len(records) + len(self) > self.size:
-            raise Exception(
-                "Index with capacity {} and {} current entries cannot add {} records".format(
-                    self.size, len(self), len(records)
-                )
-            )
-
-        for i, record in enumerate(records):
-            id = record["id"]
-            vector = record["embedding"]
-            self.id_to_seq_id[id] = record["seq_id"]
-            if id in self.deleted_ids:
-                self.deleted_ids.remove(id)
-
-            if id in self.id_to_index:
-                # Update
-                index = self.id_to_index[id]
-                self.vectors[index] = vector
-            else:
-                # Add
-                self.id_to_index[id] = self.curr_index
-                self.index_to_id[self.curr_index] = id
-                self.vectors[self.curr_index] = vector
-                self.curr_index += 1
-
-    # TODO: use id type?
-    def delete(self, records: List[EmbeddingRecord]) -> None:
-        for record in records:
-            id = record["id"]
-            if id in self.id_to_index:
-                index = self.id_to_index[id]
-                self.deleted_ids.add(id)
-                del self.id_to_index[id]
-                del self.index_to_id[index]
-                del self.id_to_seq_id[id]
-                self.vectors[index].fill(0)
-            else:
-                logger.warning(f"Delete of nonexisting embedding ID: {id}")
-
-    def has_id(self, id: str) -> bool:
-        """Returns whether the index contains the given ID"""
-        return id in self.id_to_index and id not in self.deleted_ids
-
-    def get_vectors(
-        self, ids: Optional[Sequence[str]] = None
-    ) -> Sequence[VectorEmbeddingRecord]:
-        target_ids = ids or self.id_to_index.keys()
-
-        return [
-            VectorEmbeddingRecord(
-                id=id,
-                embedding=self.vectors[self.id_to_index[id]].tolist(),
-                seq_id=self.id_to_seq_id[id],
-            )
-            for id in target_ids
-        ]
-
-    def query(self, query: VectorQuery) -> Sequence[Sequence[VectorQueryResult]]:
-        np_query = np.array(query["vectors"])
-        allowed_ids = (
-            None if query["allowed_ids"] is None else set(query["allowed_ids"])
-        )
-        distances = np.apply_along_axis(
-            lambda query: np.apply_along_axis(self.distance_fn, 1, self.vectors, query),
-            1,
-            np_query,
-        )
-
-        indices = np.argsort(distances).tolist()
-        # Filter out deleted labels
-        filtered_results = []
-        for i, index_list in enumerate(indices):
-            curr_results = []
-            for j in index_list:
-                # If the index is in the index_to_id map, then it has been added
-                if j in self.index_to_id:
-                    id = self.index_to_id[j]
-                    if id not in self.deleted_ids and (
-                        allowed_ids is None or id in allowed_ids
-                    ):
-                        curr_results.append(
-                            VectorQueryResult(
-                                id=id,
-                                distance=distances[i][j],
-                                seq_id=self.id_to_seq_id[id],
-                                embedding=self.vectors[j].tolist(),
-                            )
-                        )
-            filtered_results.append(curr_results)
-        return filtered_results
 
 
 class LocalHnswSegment(VectorReader):
@@ -520,6 +265,9 @@ class LocalHnswSegment(VectorReader):
             index = cast(hnswlib.Index, self._index)
             for i in range(len(deleted_ids)):
                 id = deleted_ids[i]
+                # Never added this id to hnsw, so we can safely ignore it for deletions
+                if id not in self._id_to_label:
+                    continue
                 label = self._id_to_label[id]
 
                 index.mark_deleted(label)
@@ -532,7 +280,7 @@ class LocalHnswSegment(VectorReader):
 
             next_label = self._total_elements_added + 1
             for i in range(len(written_ids)):
-                if batch.is_new(written_ids[i]):
+                if written_ids[i] not in self._id_to_label:
                     labels_to_write[i] = next_label
                     next_label += 1
                 else:
@@ -586,322 +334,10 @@ class LocalHnswSegment(VectorReader):
                             )
                 elif op == Operation.ADD:
                     if not label:
-                        batch.apply(record, True)
+                        batch.apply(record, False)
                     else:
                         logger.warning(f"Add of existing embedding ID: {id}")
                 elif op == Operation.UPSERT:
-                    batch.apply(record, is_add=label is None)
+                    batch.apply(record, label is not None)
 
             self._apply_batch(batch)
-
-
-class PersistentData:
-    """Stores the data and metadata needed for a PersistentLocalHnswSegment"""
-
-    dimensionality: Optional[int]
-    total_elements_added: int
-    max_seq_id: SeqId
-
-    id_to_label: Dict[str, int]
-    label_to_id: Dict[int, str]
-    id_to_seq_id: Dict[str, SeqId]
-
-    def __init__(
-        self,
-        dimensionality: Optional[int],
-        total_elements_added: int,
-        max_seq_id: int,
-        id_to_label: Dict[str, int],
-        label_to_id: Dict[int, str],
-        id_to_seq_id: Dict[str, SeqId],
-    ):
-        self.dimensionality = dimensionality
-        self.total_elements_added = total_elements_added
-        self.max_seq_id = max_seq_id
-        self.id_to_label = id_to_label
-        self.label_to_id = label_to_id
-        self.id_to_seq_id = id_to_seq_id
-
-    @staticmethod
-    def load_from_file(filename: str) -> "PersistentData":
-        """Load persistent data from a file"""
-        with open(filename, "rb") as f:
-            ret = cast(PersistentData, pickle.load(f))
-            return ret
-
-
-class PersistentLocalHnswSegment(LocalHnswSegment):
-    METADATA_FILE: str = "index_metadata.pickle"
-    # How many records to add to index at once, we do this because crossing the python/c++ boundary is expensive (for add())
-    # When records are not added to the c++ index, they are buffered in memory and served
-    # via brute force search.
-    _batch_size: int = 1000
-    _brute_force_index: Optional[BruteForceIndex]
-    _curr_batch: Batch
-    # How many records to add to index before syncing to disk
-    _sync_threshold: int = 1000
-    _persist_data: PersistentData
-    _persist_directory: str
-
-    def __init__(self, system: System, segment: Segment):
-        super().__init__(system, segment)
-        self._persist_directory = system.settings.require("persist_directory")
-        self._curr_batch = Batch()
-        self._brute_force_index = None
-        if not os.path.exists(self._get_storage_folder()):
-            os.makedirs(self._get_storage_folder())
-        # Load persist data if it exists already, otherwise create it
-        if self._index_exists():
-            self._persist_data = PersistentData.load_from_file(
-                self._get_metadata_file()
-            )
-            self._dimensionality = self._persist_data.dimensionality
-            self._total_elements_added = self._persist_data.total_elements_added
-            self._max_seq_id = self._persist_data.max_seq_id
-            self._id_to_label = self._persist_data.id_to_label
-            self._label_to_id = self._persist_data.label_to_id
-            self._id_to_seq_id = self._persist_data.id_to_seq_id
-        else:
-            self._persist_data = PersistentData(
-                self._dimensionality,
-                self._total_elements_added,
-                self._max_seq_id,
-                self._id_to_label,
-                self._label_to_id,
-                self._id_to_seq_id,
-            )
-
-    def _index_exists(self) -> bool:
-        """Check if the index exists via the metadata file"""
-        return os.path.exists(self._get_metadata_file())
-
-    def _get_metadata_file(self) -> str:
-        """Get the metadata file path"""
-        return os.path.join(self._get_storage_folder(), self.METADATA_FILE)
-
-    def _get_storage_folder(self) -> str:
-        """Get the storage folder path"""
-        folder = os.path.join(self._persist_directory, str(self._id))
-        return folder
-
-    @override
-    def _init_index(self, dimensionality: int) -> None:
-        index = hnswlib.Index(space=self._params.space, dim=dimensionality)
-        self._brute_force_index = BruteForceIndex(
-            size=self._batch_size,
-            dimensionality=dimensionality,
-            space=self._params.space,
-        )
-
-        # Check if index exists and load it if it does
-        if self._index_exists():
-            index.load_index(
-                self._get_storage_folder(),
-                is_persistent_index=True,
-                max_elements=int(
-                    max(self.count() * self._params.resize_factor, DEFAULT_CAPACITY)
-                ),
-            )
-        else:
-            index.init_index(
-                max_elements=DEFAULT_CAPACITY,
-                ef_construction=self._params.construction_ef,
-                M=self._params.M,
-                is_persistent_index=True,
-                persistence_location=self._get_storage_folder(),
-            )
-
-        index.set_ef(self._params.search_ef)
-        index.set_num_threads(self._params.num_threads)
-
-        self._index = index
-        self._dimensionality = dimensionality
-
-    def _persist(self) -> None:
-        """Persist the index and data to disk"""
-        index = cast(hnswlib.Index, self._index)
-
-        # Persist the index
-        index.persist_dirty()
-
-        # Persist the metadata
-        self._persist_data.dimensionality = self._dimensionality
-        self._persist_data.total_elements_added = self._total_elements_added
-        self._persist_data.max_seq_id = self._max_seq_id
-
-        # TODO: This should really be stored in sqlite or the index itself
-        self._persist_data.id_to_label = self._id_to_label
-        self._persist_data.label_to_id = self._label_to_id
-        self._persist_data.id_to_seq_id = self._id_to_seq_id
-
-        with open(self._get_metadata_file(), "wb") as metadata_file:
-            pickle.dump(self._persist_data, metadata_file, pickle.HIGHEST_PROTOCOL)
-
-    @override
-    def _apply_batch(self, batch: Batch) -> None:
-        super()._apply_batch(batch)
-        if (
-            self._total_elements_added - self._persist_data.total_elements_added
-            >= self._sync_threshold
-        ):
-            self._persist()
-
-    @override
-    def _write_records(self, records: Sequence[EmbeddingRecord]) -> None:
-        """Add a batch of embeddings to the index"""
-        if not self._running:
-            raise RuntimeError("Cannot add embeddings to stopped component")
-
-        # TODO: THREAD SAFETY
-        for record in records:
-            if record["embedding"] is not None:
-                self._ensure_index(len(records), len(record["embedding"]))
-
-            self._brute_force_index = cast(BruteForceIndex, self._brute_force_index)
-
-            self._max_seq_id = max(self._max_seq_id, record["seq_id"])
-            id = record["id"]
-            op = record["operation"]
-            exists_in_index = self._id_to_label.get(
-                id, None
-            ) or self._brute_force_index.has_id(id)
-
-            if op == Operation.DELETE:
-                if exists_in_index:
-                    self._curr_batch.apply(record)
-                    self._brute_force_index.delete([record])
-                else:
-                    logger.warning(f"Delete of nonexisting embedding ID: {id}")
-
-            elif op == Operation.UPDATE:
-                if record["embedding"] is not None:
-                    if exists_in_index:
-                        self._curr_batch.apply(record)
-                        self._brute_force_index.upsert([record])
-                    else:
-                        logger.warning(
-                            f"Update of nonexisting embedding ID: {record['id']}"
-                        )
-            elif op == Operation.ADD:
-                if record["embedding"] is not None:
-                    if not exists_in_index:
-                        self._curr_batch.apply(record, is_add=True)
-                        self._brute_force_index.upsert([record])
-                    else:
-                        logger.warning(f"Add of existing embedding ID: {id}")
-            elif op == Operation.UPSERT:
-                if record["embedding"] is not None:
-                    self._curr_batch.apply(record, is_add=exists_in_index is False)
-                    self._brute_force_index.upsert([record])
-
-            if len(self._curr_batch) >= self._batch_size:
-                self._apply_batch(self._curr_batch)
-                self._curr_batch = Batch()
-                self._brute_force_index.flush()
-
-    @override
-    def count(self) -> int:
-        return len(self._id_to_label) + self._curr_batch.count()
-
-    @override
-    def get_vectors(
-        self, ids: Optional[Sequence[str]] = None
-    ) -> Sequence[VectorEmbeddingRecord]:
-        """Get the embeddings from the HNSW index and layered brute force batch index"""
-        results = []
-        ids_hnsw: Set[str] = set()
-        ids_bf: Set[str] = set()
-
-        if self._index is not None:
-            ids_hnsw = set(self._id_to_label.keys())
-        if self._brute_force_index is not None:
-            ids_bf = set(self._curr_batch.get_written_ids())
-        target_ids = ids or list(ids_hnsw.union(ids_bf))
-        hnsw_labels = []
-
-        for id in target_ids:
-            if id in ids_bf:
-                self._brute_force_index = cast(BruteForceIndex, self._brute_force_index)
-                results.append(self._brute_force_index.get_vectors([id])[0])
-            elif id in ids_hnsw:
-                hnsw_labels.append(self._id_to_label[id])
-
-        if len(hnsw_labels) > 0 and self._index is not None:
-            vectors = cast(Sequence[Vector], self._index.get_items(hnsw_labels))
-
-            for label, vector in zip(hnsw_labels, vectors):
-                id = self._label_to_id[label]
-                seq_id = self._id_to_seq_id[id]
-                results.append(
-                    VectorEmbeddingRecord(id=id, seq_id=seq_id, embedding=vector)
-                )
-
-        return results
-
-    @override
-    def query_vectors(
-        self, query: VectorQuery
-    ) -> Sequence[Sequence[VectorQueryResult]]:
-        if self._index is None and self._brute_force_index is None:
-            return [[] for _ in range(len(query["vectors"]))]
-
-        k = query["k"]
-        if k > self.count():
-            logger.warning(
-                f"Number of requested results {k} is greater than number of elements in index {self.count()}, updating n_results = {self.count()}"
-            )
-            k = self.count()
-
-        # Overquery by updated elements amount because they may
-        # hide the real nearest neighbors in the hnsw index
-        hnsw_query = VectorQuery(
-            vectors=query["vectors"],
-            k=k + self._curr_batch.update_count,
-            allowed_ids=query["allowed_ids"],
-            include_embeddings=query["include_embeddings"],
-            options=query["options"],
-        )
-
-        # For each query vector, we want to take the top k results from the
-        # combined results of the brute force and hnsw index
-        results: List[List[VectorQueryResult]] = []
-        self._brute_force_index = cast(BruteForceIndex, self._brute_force_index)
-        bf_results = self._brute_force_index.query(query)
-        hnsw_results = super().query_vectors(hnsw_query)
-        for i in range(len(query["vectors"])):
-            # Merge results into a single list of size k
-            bf_pointer: int = 0
-            hnsw_pointer: int = 0
-            curr_bf_result: Sequence[VectorQueryResult] = bf_results[i]
-            curr_hnsw_result: Sequence[VectorQueryResult] = hnsw_results[i]
-            curr_results: List[VectorQueryResult] = []
-            # In the case where filters cause the number of results to be less than k,
-            # we set k to be the number of results
-            total_results = len(curr_bf_result) + len(curr_hnsw_result)
-            if total_results == 0:
-                results.append([])
-            else:
-                while len(curr_results) < min(k, total_results):
-                    if bf_pointer < len(curr_bf_result) and hnsw_pointer < len(
-                        curr_hnsw_result
-                    ):
-                        bf_dist = curr_bf_result[bf_pointer]["distance"]
-                        hnsw_dist = curr_hnsw_result[hnsw_pointer]["distance"]
-                        if bf_dist <= hnsw_dist:
-                            curr_results.append(curr_bf_result[bf_pointer])
-                            bf_pointer += 1
-                        else:
-                            curr_results.append(curr_hnsw_result[hnsw_pointer])
-                            hnsw_pointer += 1
-                    if bf_pointer >= len(curr_bf_result):
-                        remaining = k - len(curr_results)
-                        curr_results.extend(
-                            curr_hnsw_result[hnsw_pointer : hnsw_pointer + remaining]
-                        )
-                    if hnsw_pointer >= len(curr_hnsw_result):
-                        remaining = k - len(curr_results)
-                        curr_results.extend(
-                            curr_bf_result[bf_pointer : bf_pointer + remaining]
-                        )
-                results.append(curr_results)
-        return results
